pyproject.toml
[tool.poetry]
name = "fastapi-token"
version = "0.1.0"
description = ""
authors = ["Ramil Mavliutov <stirk2281337@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_token"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.97.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.16"
alembic = "^1.11.1"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^11.0.0"}
pytest = "^7.3.2"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.1"
psycopg2 = "^2.9.6"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
asyncio_mode="auto"